requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
networkx>=3.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
fashion-clip @ git+https://github.com/patrickjohncyh/fashion-clip.git 